Los anticuerpos vuelan

+++ Nanobot.h
@@ -7,30 +7,30 @@ using namespace std;
 class Nanobot : public Elemento{
     //atributos
     private:
-        int posicion_x;
-        int posicion_y;
+        float posicion_x;
+        float posicion_y;
 
     public:
         //constructores
         //sin parametros
         Nanobot();
         //con parametros
-        Nanobot(int, int);
+        Nanobot(float, float);
         //metodos
         //SETTERS
         //PRE: recibir un dato valido
         //POS: asigna posicion en x al nanobot
-        void asignar_posicion_x(int);
+        void asignar_posicion_x(float);
         //PRE: recibir un dato valido
         //POS: asigna posicion en y al nanobot
-        void asignar_posicion_y(int);
+        void asignar_posicion_y(float);
         //GETTERS
         //PRE: -
         //POS: Devuelve la posicion en x del nanobot
-        int obtener_posicion_x();
+        float obtener_posicion_x();
         //PRE: -
         //POS: Devuelve la posicion en y del nanobot
-        int obtener_posicion_y();
+        float obtener_posicion_y();
 };
